requirements.txt
python-telegram-bot==20.7
Flask==2.2.5
requests==2.31.0
